requirements.txt
paho-mqtt==1.6.1
pytest==8.3.3
